fix(models): split client helpers from server store
